+++ argopy/stores/argo_index_pa.py
@@ -1,3 +1,4 @@
+
 """
 Argo file index store