+++ argopy/utils/casting.py
@@ -177,7 +177,10 @@ def cast_Argo_variable_type(ds, overwrite=True):
         return da
 
     for v in ds.variables:
-        if overwrite or ("casted" in ds[v].attrs and ds[v].attrs["casted"] == 0):
+        if (overwrite or
+            ("casted" in ds[v].attrs and ds[v].attrs["casted"] == 0) or
+            (not overwrite and "casted" in ds[v].attrs and ds[v].attrs["casted"] == 1 and ds[v].dtype == 'O')
+        ):
             try:
                 attrs = deepcopy(ds[v].attrs)
                 ds[v] = cast_this_da(ds[v], v)